+++ projects/ngxi4-dynamic-service/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-# 2.0.9 add delay for reference 
+# 3.0.2 add "ng-multiselect-dropdown": "^0.2.10" 
+
+# 2.0.9 & 2.1.0 add delay for reference 
 
 # 2.0.8 isMobile for with of window in platform